librbd: move read callback helpers to narrowest scope
(cherry picked from commit 65b336f68596a430629692a8682c5cfe883c31f1)
